adiciona o ambiente de testes e o teste de soma
