+++ package.json
@@ -1,5 +1,10 @@
 {
   "dependencies": {
-    "express": "^4.18.1"
-  }
+    "express": "^4.18.1",
+    "jest": "^28.1.3",
+    "supertest": "^6.2.4"
+  },
+    "scripts": {
+      "test": "jest"
+    }
 }
